requirements.txt

torch>=2.2
torchaudio>=2.2
pytorch-lightning>=2.5.1.post0
librosa>=0.10
numpy
scipy
tqdm
pyyaml
matplotlib
scikit-learn
torchmetrics>=1.3
streamlit
torchvision>=0.15.0
